update VSIX to be installed by Willow
We need to make it a product component and set "AllUsers" so we can use a known installation path.

I'm renaming the known path to "Xamarin.Templates" to better match the content of the new VSIX. I've tried a VSIX upgrade from path X to path Y in a newer VSIX and it does work ok.