Merge 15.5 into master (#94)
* Fixes Azure Dialog Height in order to avoid cutting of the `UWP` option (#46)

* Update MainPage.xaml

* Fixed wrong iOS version for 11.0
When the version number is 11.0, it was wrongly set as 11 (without the
trialing 0). Apparently this is because at some point the value is
converted to a number and the decimals discarded. Adding quotes prevents
this behavior. This fixes
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/510337

* Added Localization to Xamarin Templates
Added new packages for localization plus some changes in targets for
them to work correctly

* Add symbol for Xamarin.Forms version.

* Added proper naming for insertion (#62)

* Create macOS project.

* update VSIX to be installed by Willow

We need to make it a product component and set "AllUsers" so we can use a known installation path.

I'm renaming the known path to "Xamarin.Templates" to better match the content of the new VSIX. I've tried a VSIX upgrade from path X to path Y in a newer VSIX and it does work ok.

* add PackageId to match legacy Willow package

* Fixed sln folder location
Template was unfolding on App\App, with sln files in both directories.
This moves the structure to be just App, and directories for the
platforms and shared projects inside that. sln will still be on the top
level folder, but not duplicated. This is similar to what
duplo-templates did.

* Fix #510342. (#77)

* Update xamarincrossplatformapp.vstemplate

* Bump Forms to 2.5.0.77107.

* Update xamarincrossplatformapp.vstemplate

* Update template.json

* Update template.json

* Update PULL_REQUEST_TEMPLATE.md

* Update GitInfo.txt

* Update Forms version to latest (#89)

Manually doing this due a bug in the templating engine

* bump version