Fix Xamarin.Forms master-details build errors and warnings (#162)
* Fix Xamarin.Forms Android master-detail template build error.

Fixed VSTS #572135 - Build error for building Forms App - Android -
Shared project with Add ASP.NET core checkbox enabled.
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/572135

The build error:

CS0246: The type or namespace name 'HttpClient' could not be found

This was because when the shared project selected the Android project
needed to reference System.Net.Http so the HttpClient could be found.

* Fix Xamarin.Forms master-detail UI Test build warnings

Fixed VSTS #572146 - Getting Reference warning for iOS after building
the Forms App - Android - Shared project with UI test enabled.
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/572146

When only the Android or iOS project is created on their own the
UITest project would still have project references to both projects,
one of which would be missing. Now the project references are added
only if the corresponding Android or iOS project would be created.