Fix Forms template build error when Xamarin.Essentials not used
The AzureDataStore classes use Xamarin.Essentials which is optional
so the code does not compile if Xamarin.Essentials is not used.
The code is now conditional based on whether Xamarin.Essentials
is being used.

Fixes VSTS #669955 - Xamarin.Forms Tabbed App fails to build with
The type or namespace name 'Essentials' does not exist in the
namespace 'Xamarin' (are you missing an assembly reference?) (CS0234)
in AzureDataStore.cs
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/669955