Update cross-platform FallbackAndroidLevel
I tested this change locally, and it had the desired effect.  Note that
the fallback mechanism for this template doesn't lead to a project that
is strictly correct.  The Xamarin.Forms package is *not* affected by
this logic, so it will almost certainly depend on a higher Android
Support Library version than is installed in the fallback scenario.
This raises a question about whether it would be better to constrain the
Xamarin.Forms package version across *all* of the projects (iOS, UWP,
and .NET Standard library) based on the Support Library version selected
for the Xamarin.Android project.  That doesn't necessarily seem like a
good strategy because users really might be trying to create a solution
that uses that latest Xamarin.Forms version.  From their perspective, it
might just be "accidental" that they don't have the right Android API
level installed to allow all the latest NuGet packages.  Long story
short, I have left that rough edge alone for now and just updated the
FallbackAndroidLevel to match the non-cross-platform Android project.