Enable remote settings to control opening of XAML and CS
Two new remote settings with a boolean control the opening
of the `.xaml` and `.xaml.cs` files: `Xamarin.OpenXaml` and
`Xamarin.OpenXamlCs`.

This will allow us to set up an A/B experiment to verify the
impact of both actions in terms of performance and retention.

To test this behavior:

1. Download the [LocalRemoteSetting.ps1](https://mseng.visualstudio.com/VS%20IDE%20-%20Telemetry/_git/Tools?path=%2FRemoteSettings%2FLocalRemoteSetting.ps1&version=GBmaster&_a=contents)
   and unblock it using Windows explorer file properties window.
2. From powershell, run a command like the following: `.\LocalRemoteSetting.ps1 -collectionPath Xamarin -propertyName OpenXaml -propertyValue false` or `.\LocalRemoteSetting.ps1 -collectionPath Xamarin -propertyName OpenXamlCs -propertyValue false`