Change default of OpenXaml/OpenXamlCs settings to false
As discussed with Alan, he's not opposed to not opening the
files during solution creation. The mechanism of enabling it
via A/B testing is a better way of gauging this impact before
we enable it again, given we know for sure it causes significant
performance impact due to the designer infrastructure having
to initialize very early in the process.