Add an MSI-based NuGet cache
The MSI contains the same contents as the former VSIX
cache, but it additionally drops a NuGet configuration
file in C:\Program Files (x86)\NuGet\Config so it gets
registered as a fallback folder for packages.

For first run scenarios, this should significantly
improve our numbers. In my testing, after running
`nuget locals all -clear`, the before/after this MSI
is installed is about ~10 seconds delta with a 50Mbps
internet connection (since packages need to be downloaded).
It will obviously be way more significant for customers
with slow bandwidth.