Initialize packages while the UI thread is idle waiting for user input
At the point we do `ShowDialog()`, the user will be presented with the UI and
will have to analyze the options and make a decision before clicking Create.
This will usually take a couple seconds, during which the UI thread will be
idle, at which point we can initialize our packages essentially for free.

This is similar to what the designer is doing, so we switch to using the new
`IVsShell7` API too to load the packages asynchronously. By doing each in its
own `StartOnIdle`, we let the UI do work in between those different package
load, to avoid UI delays that might happen otherwise if they take some time
as a whole.