Moving NuGet packages off the System Drive
The Setup team requested us to support the installation of our NuGet packages cache off the system drive. This is a big push the Setup team is doing across all of VS in order to move bulky bits to secondary, cheaper storage disks (usually HDD) instead of system drive (usually SDD).

Given that we have a few NuGet packages that do need to write files to their install location we cannot use the provided [SharedInstallDir] as we need to make sure we install to a non-UAC protected folder.

Other SDKs like Android NDK and SDK do install into a root folder named  just "Microsoft". We will follow that pattern here for our bits and will install to:

[SharedInstallDrive]\Microsoft\Xamarin\NuGet

This [SharedInstallDrive] will be passed from Willow into our MSI via our new property 'CACHEDRIVE'.

Our config files still needs to be copied to a well-known location where NuGet can find it, so we are not touching that file with this change.